Implemented end to end integration test
Currently, we've not yet mocked out the upstream terraform registry due
to difficulties with managing the zip file. Also, the port of the test
server for this test case is currently fixed due to a hard requirement
to use HTTPS by the terraform client. We'll double back to solve this at
a later time.