ci: Use AWS provider in integration test
Turns out the two providers I was using was not large enough to trigger
an edge case in artifacts.rs due to their small size as they fit in a
single multi part upload. I've swapped out one of the providers for the
AWS provider to test "large" streaming uploads.