feat!: Move location of terraform provider endpoint
This is being done to segregate mirror endpoints from other endpoints
such as API endpoints.
We are doing this early in the project to prevent issues with
namespacing in the future (i.e. imagine a hostname being hosted at `api`
on an internal network).
To allow users to migrate from the old location, we will make one
release to simultaneously support the endpoints at the old location which could be
deployed as a stepping stone release so no downtime is incurred from a
version upgrade.