chore: Removed experimental lambda runmode
I've decided to not support running terrashine from a lambda function.
This decision is due to lack of progress and difficulty in scheduling
the refresh task without bringing in additional components like SQS for
background jobs. To keep the application lean, we'll only support
running this as an old-school server application. I'll likely start
looking at centralized job scheduling inside of postgresSQL after this.